# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZENOVA Platform API",
    "description": "Research-Grade Mental-Health Wellbeing and Conversational Support Platform API",
    "contact": {
      "name": "ZENOVA Engineering & Clinical Team",
      "url": "https://zenova.ai/",
      "email": "support@zenova.ai"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/escalations/rules": {
      "get": {
        "tags": [
          "Escalations & Clinician Dashboard"
        ],
        "summary": "List Configured Clinician Rules",
        "description": "Inspect documented clinician rules and trigger configurations.",
        "operationId": "list_configured_clinician_rules_api_v1_escalations_rules_get",
        "parameters": [
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician-admin-1",
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Configured Clinician Rules Api V1 Escalations Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalations": {
      "get": {
        "tags": [
          "Escalations & Clinician Dashboard"
        ],
        "summary": "List Escalations",
        "description": "List crisis escalation alert tickets with role-based filtering and data masking.",
        "operationId": "list_escalations_api_v1_escalations_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "clinician_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Clinician Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician-admin-1",
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalations/{alert_id}": {
      "get": {
        "tags": [
          "Escalations & Clinician Dashboard"
        ],
        "summary": "Get Escalation Alert",
        "description": "Retrieve detailed clinical alert with context snapshot (masked for non-clinicians).",
        "operationId": "get_escalation_alert_api_v1_escalations__alert_id__get",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician-admin-1",
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Escalation Alert Api V1 Escalations  Alert Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalations/trigger": {
      "post": {
        "tags": [
          "Escalations & Clinician Dashboard"
        ],
        "summary": "Trigger Escalation Alert",
        "description": "Programmatically trigger and record a new clinical alert.",
        "operationId": "trigger_escalation_alert_api_v1_escalations_trigger_post",
        "parameters": [
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician-admin-1",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Trigger Escalation Alert Api V1 Escalations Trigger Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalations/{alert_id}/acknowledge": {
      "post": {
        "tags": [
          "Escalations & Clinician Dashboard"
        ],
        "summary": "Acknowledge Escalation",
        "description": "Mark an escalation alert as claimed and acknowledged by a clinician.",
        "operationId": "acknowledge_escalation_api_v1_escalations__alert_id__acknowledge_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician-admin-1",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeAlertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalations/{alert_id}/status": {
      "post": {
        "tags": [
          "Escalations & Clinician Dashboard"
        ],
        "summary": "Update Alert Status",
        "description": "Transition alert lifecycle state (e.g. in_review or dismissed).",
        "operationId": "update_alert_status_api_v1_escalations__alert_id__status_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician-admin-1",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalations/{alert_id}/resolve": {
      "post": {
        "tags": [
          "Escalations & Clinician Dashboard"
        ],
        "summary": "Resolve Escalation",
        "description": "Resolve an alert with mandatory action taken and clinical justification notes.",
        "operationId": "resolve_escalation_api_v1_escalations__alert_id__resolve_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician-admin-1",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveAlertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/escalations/{alert_id}/audit": {
      "get": {
        "tags": [
          "Escalations & Clinician Dashboard"
        ],
        "summary": "Get Alert Audit History",
        "description": "Retrieve immutable forensic audit trail for an escalation ticket.",
        "operationId": "get_alert_audit_history_api_v1_escalations__alert_id__audit_get",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "clinician-admin-1",
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Alert Audit History Api V1 Escalations  Alert Id  Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AcknowledgeAlertRequest": {
        "properties": {
          "clinician_id": {
            "type": "string",
            "title": "Clinician Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "clinician_id"
        ],
        "title": "AcknowledgeAlertRequest"
      },
...
      "AlertListResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "alerts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Alerts"
          }
        },
        "type": "object",
        "required": [
          "count",
          "alerts"
        ],
        "title": "AlertListResponse"
      },
...
      "CreateAlertRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "severity": {
            "$ref": "#/components/schemas/EscalationSeverity"
          },
          "trigger_type": {
            "$ref": "#/components/schemas/EscalationTriggerType"
          },
          "reason_code": {
            "type": "string",
            "title": "Reason Code"
          },
          "reason_title": {
            "type": "string",
            "title": "Reason Title"
          },
          "reason_description": {
            "type": "string",
            "title": "Reason Description"
          },
          "trigger_cues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trigger Cues"
          },
          "user_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Message",
            "default": ""
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "user_id",
          "severity",
          "trigger_type",
          "reason_code",
          "reason_title",
          "reason_description"
        ],
        "title": "CreateAlertRequest"
      },
...
      "EscalationSeverity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "EscalationSeverity",
        "description": "Urgency level of the clinical alert."
      },
...
      "EscalationTriggerType": {
        "type": "string",
        "enum": [
          "crisis_risk",
          "repeated_signals",
          "longitudinal_change",
          "safety_gate_event",
          "clinician_rule"
        ],
        "title": "EscalationTriggerType",
        "description": "Origin category that caused the escalation trigger."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  },
...
}